+++ CHANGELOG.md
@@ -1,10 +1,20 @@
 # Changelog
 
 
+### 0.0.6 (2018-07-14)
+
+- major code refactoring
+- even better error catching mechanism
+- added titles to user settings (checkboxes)
+- added timeout feature (60 seconds) and "Timed out" UI status
+- various fixes
+
+
+
 ### 0.0.5 (2018-07-11)
 
 - remove jQuery dependency
-- better catch mechanism
+- better error catching mechanism
 - various fixes